Refactored bradley_2013
